refactor(components): import 경로 변경 ('..' > '@/components')
